codec: redesign how codecFnInfo is loaded from, and support SelfExt
Having to redesign your values to support extensions is a flawed design.
We have now reverted that, and allow users to just signify within the current
design that they want to support the native encoding/decoding for their extensions.

However, we needed to fix some flaws in the way codecFn was loaded previously.
That (my original) design is flawed and needs to be fixed before we build
anything else atop it.

For one, the fact that we pass checkFastpath and checkCodecSelfer were
code smells, that something wasn't right with the design. Fundamentally,
the fn() should be a static function that always returns the same value
for a given context (since it caches them). I need to just re-architect
it and remove those functions.

Finally, codec.Selfer can be an extension.
Consequently, it cannot delegate to EncodeExtension itself.
Ensure that the check for extension happens before the check whether a Selfer or not.

Summary of changes:

- BasicHandle now keeps 2 rtidFns: one for all, and one for extensions that encode natively
- fn() no longer takes checkFastpath or checkCodecSelfer
- formats all support the sentinel SelfExt
- ext encoding is more efficient as it tries to reuse pools of []byte

Thanks @kostko for his gentle and committed push!

Fixes #274