codec: streamline unsafe (and safe) map range implementation
Now, unsafe map range is optimized for all versions of go 1.7+

We leverage some unsafe commonality that is not limited to
what is exposed in reflect package.

go 1.4+ support hashiter with the first 2 elements being
key and value, so we depend on that.

Furthermore, for safe mode, we just delegate directly to the standard
library methods and do not "cache" the results.
We know that the mapAddressableRV for safe mode is always a zero value.