codec: json: use sync.pool for decoding buffers and optimize DecodeBytes
The buffer used for decoding may be small in some cases.

We now leverage the sync.pool for buffers, and grab the buffer
we need, and return to pool when done.

Also, when decoding bytes written as base64, we do not expect a string
quotation mark (") in the string, so we just look for the ending "
and return that as the bytes in the stream.