codec: json: jsonEncDriver is one type, not a smartly assembled type
Previously, jsonEncDriver was either jsonEncDriverTypical or
jsonEncDriverGeneric based on if we needed indenting or other flags.

The elision of that branch was not a big hamper to performance
because branch prediction already made sure that we took right
branch most of the time.

Benchmarks confirm that we don't lose much from it.

So - consolidate into one type.

This also allows further optimizations e.g. making Encoder a
non-pointer field of jsonEncDriver.