codec: move Enc/Decoder into enc/decDriver for better memory arrangement
We reduce some pointer chasing by having the Encoder and Decoder be part of the
drivers respectively.

This should allow the referencing of the decoder/encoder from the driver without
dereferencing/loading some other value from memory.

This required some re-architecture of the data structures to support the usage model.
For example, New Encoder(writer, handle) now does:

    var e = handle.newEncDriver().encoder()
    e.Reset(w)
    return e

It resulted in some but not much of a performance difference, but I like the new
usage model and data structures better.