codec: StringToRaw should only be checked in the Handle
Previously, we checked StringToRaw everytime a string is to be encoded.
This helped prevent an unnecessary function call overhead, but also
exploded the size of generated files and users had to explicitly check
each time.

Now, we put it into the Handle, so the function call overhead can be
eliminated at the Handle's encDriver implementation.