codec: remove EncodeAsis and optimize access to jsonEncDriver
First, EncodeAsis is only supported by jsonEncDriver and does the default
implementation of writing bytes to the stream as-is.
There's no need to have an encDriverAsis interface.

So - remove all traces of encDriverAsis, EncodeAsis, etc.

Secondly, we do not need to keep a reference to jsonEncDriver and jsonDecDriver
as those can be retrieved quite performantly via encDriver and decDriver interface.
Instead, we make safe/unsafe ways to grab it easily.

Finally, create the option for making optimized checkBreak that affords
inlining. This is not an option right now, as CheckBreak cannot be inlined.
But set things in place for it.