codec: json: optimize reading escaped string and inline decRd.(un)readn1, encWr.write(n2|qstr)
Previously, reading escaped strings involved reading up to ", and then
parsing it in a loop. This means that we iterated the strings twice.

This is now optimized by just reading one char at a time from the stream.

There is a performance cost when using IO streams, as each read involves a function
call. However, this simplifies the code, and makes the typical path of reading
from a []byte faster.

To accomplish this, we needed to ensure that decRd.(un)readn1 is inlined.
We achieved that.

While at it, we also ensured that encWr.write(n2|qstr) are inlined.

One change is that we no longer process incomplete unicode strings by replacing
with the replacement char. Instead, if an EOF is reached early, we just return an
io.EOF error.

This change makes other codes faster because those inlined methods are the most
commonly used.