codec: support ZeroCopy for allowing decoded output point into input bytes
ZeroCopy controls whether decoded values point into the
input bytes passed into a NewDecoderBytes/ResetBytes(...) call.

To illustrate, if ZeroCopy and decoding from a []byte (not io.Writer),
then a []byte in the output result may just be a slice of (point into)
the input bytes.

This optimization prevents unnecessary copying.

However, it is made optional, as the caller MUST ensure that the input parameter
is not modified after the Decode() happens.

Fixes #312