codec: use binary search to find index into sorted array
Previously, we embedded all field names and lengths in a []byte
and used bytes.Index(...) to find it.
However, bytes.Index is expensive and cannot use the lengths of the
strings as a quick optimization.

Doing this way also allows indexForEncName to be inlined, and it shows
highly on benchmarks.