codec: optimize mapRange/iteration and mapGet
The only callers of mapRange/iteration is encode. Here, we just walk through the values
and encode them

The only caller of mapGet is decode. Here, it does a Get if the underlying value is a pointer,
and decodes into that.

For both users, we are very careful NOT to modify or keep the pointers around.
Consequently, it is ok for take advantage of the performance that the map is not modified
during an iteration and we can just "peek" at the internal value" in the map and use it.