codec: optimize nextValueBytes so it does not allocate when not needed.
There are 2 instances where allocation is not necessary
- when decoding from a []byte input. We can just return a sub-slice
- on swallow, when we just want to discard the next value from the stream

In both instances, we can optimize appropriately.

If a caller needs to use the bytes, pass in a non-nil slice.
Passing a nil slice says that we do not care about the returned bytes.

Callers that need to re-use the bytes in a case where returned byte is a
sub-slice of the input, can check d.bytes and d.h.ZeroCopy to determine
whether a copy is needed e.g. Decoder.rawBytes.

Other changes:
- decoding arrays should use fast-path for slices if available
- remove seqType, as its not needed.