codec: robust handling of codecgen caveats wrt Canonical|CheckCircularRef flags and MissingFielder implementations
- Canonical option.
  If Canonical=true, codecgen'ed code will delegate encoding maps to reflection-based code.
- MissingFielder implementation.
  If a type implements MissingFielder, a Selfer is not generated (with a warning message).
- CheckCircularRef.
  When encoding a struct, a circular reference can lead to a stack overflow.
  If CheckCircularRef=true, codecgen'ed code will delegate encoding structs to reflection-based code.

To do this, we add a sentinel method to each generated file, so we can tell if the type was generated by codecgen.
We use this when handling CheckCircularRef to determine whether codecgen created the type, and if so, we should
not use the Selfer implementations.

We also now log warnings if attempting to generate for a type that implements MissingFielder.

Finally, we look at Canonical flag and delegate to reflection based code when encoding maps.

Also, updated genVersion to 24.