codec: remove atEndOf(En|De)code calls
only supported by binc and json:
- binc: encode/decode: clear symbol maps
- json: encode only:   append ' ' or '\n' so that numbers are not concatenated when encoded

Only json needs to support this. For binc, the symbols should persist for the lifetime of the
Encoder/Decoder, so no need to clear it until a Reset is done.

Instead, handle json encode use-case as a unique json requirement honored by the Encoder.