codec: optimize transient handling in safe and unsafe modes
in safe mode: only use transient for numbers, bool, string and slice

in unsafe mode: use transient for numbers, bool, string and slice,
and for structs and arrays that have no pointers inside them.
This means that array elements and struct fields only have numbers and bool in them.

This way, we honor the contract that the memory space used for transient values
will never have any pointers and never confuse the GC.

Current implementation is much cleaner.