codec: reduce size of structFieldInfos in default mode
Currently, structFieldInfo is provided by a typeInfo in sorted and source modes.
Both are equal length slices.

In default (unsafe) mode, we can reduce the space taken from 6 words (2 slices)
to 3 words (2 pointers and a length).

This will bring size of typeInfo to 16 words, fitting it neatly in cache lines.