codec: decode a number from any number in the stream
Previously, we always decoded like types ie
- from a signed integer in the stream to an intXXX
- from an unsigned integer in the stream to an uintXXX
- from a floating point value in the stream to an floatXXX

In addition, we sometimes supported others on a format-by-format
basis e.g. we typically supported int->float but not float->int for some formats.

Moving forward, we will make it consistent. Any number can be decoded from
any number representation in the stream, with best effort and proper
error handling.

Some directions will cause an error e.g. the following are invalid:
- decode a negative integer or floating point value in the stream into an unsigned value
- decode a floating point value with a fraction into an integer value

To support this robustly, we had to do some refactoring of the codebase
- support decNegintPosintFloatNumber interface for cbor, binc, simple
  which can define how to decode an integer or a float separately,
  and a helper for transparent number handling for those formats.
- msgpack doesn't fit these formats, so we handle it in msgpack directly
- for json, we refactor decimal handling to support transparent number handling

Fixes #358