codec: encoding in canonical mode should not handle custom types as predeclared
Previously, during canonical map encoding, we treated each key as if it was
a predeclared type if the kind was numeric, bool or string. However, custom
types (e.g. type A string) which may have a custom extension or otherwise will
not have those functions called.

We've now fixed this to check if type is predeclared before we short-circuit the
full encodeValue() call.