codecgen: switch genStructCanonical=true and handle map and struct encoding smartly
During canonical encoding of maps, if the keys are predeclared type (numbers, string),
then sort the keys first, and encode them directly. Previously, we always punted
to reflection based code, which tends to be slower.

During canonical encoding of structs, make sure we track the right position
so that the ti.sfi.sorted (structfieldinfos sorted by encName)
and genFQN (slice holding fully qualified names during codecgen) will be in sync.

Turn on genStructCanonical so we generate the different code based on whether the
runtime flag Canonical=true or false.