Use base32 instead of base64 for identifiers (#21)
Base64 in Go 1.22 requires that the list of characters
in the encoding be unique. This is not possible
if we are using it to generate Go identifiers.

So, we switch to base32.

This is not a breaking change as this only affects
the generated code, and not the wire format.