codec: rearrange the structs to reduce GC pressure (per GC guide)
Per GC guide, it's best to put pointers and values containing pointers
(e.g. interfaces, some slices of pointers, fields that are values containing pointers, etc)
together around the top of a struct.

This reduces GC pressure, since a value will only be scanned up to the last
pointer containing part of it.