codec: use sync.Pool and shared encoders for side encoding/decoding
Previously, we had a setup based on type parameters/generics.
It was convoluted.

However, most encoding/decoding functions are large enough that
they cannot be inlined. The only thing saved via monomorphization
is a pointer dereference. That's not worth it.

This makes the code much simpler and still performant,
without making each caller pay the price of holding onto another encoder/decoder.