codec: use a trie to search for struct fields during decoding (not a map[string]XXX)
Previously, a map[string]XXX was used.

A trie is more performant especially for this limited case
of finding a struct field mapped to a string name.

To make it performant, each node for a character takes
only 2 words or 4 words, depending on if using optimized unsafe or safe mode.

To help with cache-coherence, we try to allocate all the nodes for the trie
from a single contiguous block of memory (see uint8To32TrieNodeArena).

We also use a sync.Pool to allow efficient re-use of the scratch tries made initially.

We see improved performance (~3-5%) based on this optimization.