codec: streamline handling errors and panics
Basically, errors are handled via panics within the library
- (En|De)code will recover a panic, and deliver an error
- Must(En|De)code will recover a panic, embellish the error with context (as a codec.Error) and re-raise it
- must(En|De)code will not try to recover a panic

To use a single method across the board, we use a sentinel value to signify
that we should recover within the function. This allows a calling function to
handle recovering while the conversion to an error or a re-raise is still handled elsewhere.

Furthermore, we introduce the use of errors.Is to handle trapping errors like io.EOF
which we have wrapped