codec: test: start of determining what to compare against based on Handle options
For example, when you encode a nil slice, if NilToZeroLength option is set, then
you must encode as an empty slice.

When decoded, you get a non-nil side on the other end. To compare to the original,
you need to clone a copy of the original with similar changes made.

This incorporates this logic into the testEqual function.