codec: json: support TimeFormat and BytesFormat config options
To support the ability within json to encode time and []byte in multiple ways,
and to decode arbitrary values regardless of how uniform they are in the stream,
we now support configuring TimeFormat []string and BytesFormat []string in JsonHandle.

To enable this, we retire jsonEncState, jsonDecState.
Instead, we create jsonHandleOpts that can be reset from JsonHandle.

jsonHandleOpts fields are used at runtime to control encoding and decoding of time.Time and []byte.